requirements.txt
# Core dependencies - only what's needed for the Folio app
# Note: Developer Dependencies in requirements-dev.txt
pandas==2.2.1
numpy==1.26.4
# scipy removed - no longer needed after migrating to QuantLib
QuantLib>=1.30  # For option calculations

# Utilities
requests>=2.32.0
PyYAML==6.0.1

# Data source
yfinance>=0.2.37  # For portfolio beta calculation

# Web application - using latest versions for security updates
dash>=2.14.2
dash-bootstrap-components>=1.5.0
dash-bootstrap-templates>=1.1.1  # For Plotly figure templates

# WSGI server for production - always use latest for security
gunicorn>=21.2.0

# AI/ML dependencies
google-generativeai>=0.3.0  # For Gemini AI integration

